Update patch set 4
Patch Set 4:

I go with Humberto here, now a build ends up with a combined boot.img and a recovery.img for those who prefer to use the separate recovery partition.

But as far as I am concerned we should end up with a separate boot.img and recovery.img.
I think when all is up and running, apply this patch and write a good and conclusive "how to install on your eagle tut" upon publishing rom

Patch-set: 4
Label: Verified=0